+++ src/estimagic/optimization/quadratic_subsolvers.py
